Probably finished the refactor
